refactor: extract zene-workspace crate (Phase 3 decoupling)
Move agent instructions, workspace overview, git branch, and skills
discovery behind WorkspaceProvider with FsWorkspaceProvider for local FS.

Pure build_system_prompt composes sections from a provider; AgentBuilder
uses FsWorkspaceProvider. Removes core workspace.rs and skills.rs.
