+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
     "crates/mcp",
     "crates/permission",
     "crates/tool-runtime",
+    "crates/workspace",
 ]
 resolver = "2"
 
@@ -55,6 +56,7 @@ zene-tools = { path = "crates/tools" }
 zene-mcp = { path = "crates/mcp" }
 zene-permission = { version = "0.1.11", path = "crates/permission" }
 zene-tool-runtime = { path = "crates/tool-runtime" }
+zene-workspace = { path = "crates/workspace" }
 unigateway-sdk = { version = "2.14", default-features = false, features = ["core"] }
 llm_providers = "0.12.0"
 tiktoken-rs = "0.12"
