+++ crates/core/Cargo.toml
@@ -27,6 +27,7 @@ zene-tools = { workspace = true }
 zene-mcp = { workspace = true }
 zene-permission = { workspace = true }
 zene-tool-runtime = { workspace = true }
+zene-workspace = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3"
