+++ crates/workspace/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "zene-workspace"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+description = "Workspace context and skills discovery for Zene agents"
+readme = "README.md"
+repository = "https://github.com/ParaTensor/zene"
+keywords = ["agent", "workspace"]
+categories = ["development-tools"]
+publish = false
+
+[dependencies]
+tracing.workspace = true
+
+[dev-dependencies]
+tempfile = "3"
