+++ crates/workspace/README.md
@@ -1,0 +1,9 @@
+# zene-workspace
+
+`WorkspaceProvider` trait and filesystem adapter for project instructions, directory overview, and skills discovery.
+
+Pure `build_system_prompt` composes sections from a provider; IO stays in `FsWorkspaceProvider`.
+
+## License
+
+MIT
